third_party/blink/web_tests/external/wpt/html/editing/dnd/datastore/datatransferitemlist-remove-expected.txt

This is a testharness.js-based test.
[FAIL] remove()ing an item will put the associated DataTransferItem object in the disabled mode
  assert_equals: expected "" but got "file"
Harness: the test ran to completion.
